requirements.txt
absl-py==2.1.0
addict==2.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
boto3==1.36.10
botocore==1.36.10
Brotli @ file:///croot/brotli-split_1736182456865/work
certifi @ file:///croot/certifi_1734473278428/work/certifi
cffi==1.17.1
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
crcmod==1.7
cryptography==44.0.0
cycler==0.12.1
Cython==3.0.11
docker-pycreds==0.4.0
easydict==1.13
filelock==3.14.0
fonttools==4.55.8
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
gmpy2 @ file:///croot/gmpy2_1738085463648/work
grpcio==1.70.0
huggingface-hub==0.28.1
idna @ file:///croot/idna_1714398848350/work
imageio==2.37.0
Jinja2 @ file:///croot/jinja2_1737760107953/work
jmespath==0.10.0
jpeg4py==0.1.4
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
lazy_loader==0.4
lmdb==1.6.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe @ file:///croot/markupsafe_1736366905094/work
matplotlib==3.10.0
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mmcv-full==1.7.2
model-index==0.1.11
mpmath @ file:///croot/mpmath_1690848262763/work
networkx @ file:///croot/networkx_1737039604450/work
numpy==1.26.4
opencv-python==4.11.0.86
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
ordered-set==4.1.0
oss2==2.17.0
packaging==24.2
pandas==2.2.1
pillow @ file:///croot/pillow_1738010226202/work
platformdirs==4.3.6
protobuf==5.29.3
psutil==6.1.1
pycocotools==2.0.8
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
PySocks @ file:///work/ci_py311/pysocks_1676822712504/work
python-dateutil==2.9.0.post0
pytorch-pretrained-bert==0.6.2
pytz==2023.4
PyYAML @ file:///croot/pyyaml_1728657952215/work
regex==2024.11.6
requests==2.28.2
rich==13.4.2
s3transfer==0.11.2
safetensors==0.5.2
scikit-image==0.25.1
scipy==1.15.1
sentry-sdk==2.20.0
setproctitle==1.3.4
six==1.17.0
smmap==5.0.2
sympy @ file:///croot/sympy_1738108488918/work
tabulate==0.9.0
tb-nightly==2.19.0a20250131
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
tifffile==2025.1.10
tikzplotlib==0.10.1
timm==1.0.14
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tornado==6.4.2
tqdm==4.65.2
triton==2.1.0
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tzdata==2025.1
urllib3==1.26.20
visdom==0.2.4
wandb==0.19.5
webcolors==24.11.1
websocket-client==1.8.0
Werkzeug==3.1.3
yacs==0.1.8
yapf==0.43.0
